adding venture carousel section o'and cleaning up unused imports
